Se modifica proyecto de prueba por proyecto gexpblog se hacen pruebas y funciona correctamente
