+++ .env
@@ -5,21 +5,21 @@ NEXT_PUBLIC_EMAILJS_USER_ID=aKiBVQeuq7g_KP6fd
 
 HOST=0.0.0.0
 PORT=1337
-APP_KEYS=7spVAo4F5deLnFM82Fyz7g==,LUE44jfCIeAoWYnr0LtaCg==,e7zweVhrDnjjf1hIXRNQiw==,mVijvb1u9PDhAsjLx9F8fg==
-API_TOKEN_SALT=oD4GXInSBT/w4klya/jn2A==
-ADMIN_JWT_SECRET=9GfrhANyOxpn+UZLw/m0Ag==
-TRANSFER_TOKEN_SALT=qm77oJacEe6OG7TIM8S4Yw==
+APP_KEYS=PtkuL0QmhcIS3bUcEBPVKA==,xaL5a2/feTWSnnBKQjWh4g==,gwotW+1+Ut/ANOzrgcNk9A==,sclfxALYQFWO678icEbiEA==
+API_TOKEN_SALT=TpjhiQJITDXN2A+hWgp4Zw==
+ADMIN_JWT_SECRET=aIg6L8fcnH6p0W2R77CxWA==
+TRANSFER_TOKEN_SALT=aHxjaPi7yLEGkMnzgvERyA==
 # Database
 DATABASE_CLIENT=postgres
-DATABASE_HOST=dpg-cgvhkn0dh87jokru4rr0-a.oregon-postgres.render.com
+DATABASE_HOST=dpg-chbd2eu7avjcvo5k1fc0-a.oregon-postgres.render.com
 DATABASE_PORT=5432
-DATABASE_NAME=guitarla_orut
+DATABASE_NAME=blog_8j59
 DATABASE_USERNAME=root
-DATABASE_PASSWORD=k4i4adJ4Je3zq8nexZtJZdYdyO1NTDvp
+DATABASE_PASSWORD=PahQ176NxMjS59AYCKBaRsPVXOqw0SNA
 DATABASE_SSL=true
-JWT_SECRET=4k25J2j2bzoUmcTTA7cF8Q==
+JWT_SECRET=eX3IilUetf3x6E+tArAe0g==
 
 NEXT_PUBLIC_API_URL=http://localhost:1337/api
-NEXT_PUBLIC_CLOUDINARY_NAME=dshsbjvtl
-CNEXT_PUBLIC_LOUDINARY_KEY=218155731198731
-NEXT_PUBLIC_CLOUDINARY_SECRET=Z7tDVYYFy3wSdO6JKjIwp6bPP0g
+NEXT_PUBLIC_CLOUDINARY_NAME=daqgtrvct
+CNEXT_PUBLIC_LOUDINARY_KEY=684496963651348
+NEXT_PUBLIC_CLOUDINARY_SECRET=Bqg8QAXwzWedi5kipWuvJ7SFaZ8
