+++ gexpblog/src/api/blog/controllers/blog.ts
@@ -1,0 +1,7 @@
+/**
+ * blog controller
+ */
+
+import { factories } from '@strapi/strapi'
+
+export default factories.createCoreController('api::blog.blog');
