+++ gexpblog/src/api/blog/routes/blog.ts
@@ -1,0 +1,7 @@
+/**
+ * blog router
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreRouter('api::blog.blog');
