+++ gexpblog/src/api/blog/services/blog.ts
@@ -1,0 +1,7 @@
+/**
+ * blog service
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreService('api::blog.blog');
