--- strapi/guitarla-strapi/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "guitarla-strapi",
-  "private": true,
-  "version": "0.1.0",
-  "description": "A Strapi application",
-  "scripts": {
-    "develop": "strapi develop",
-    "start": "strapi start",
-    "build": "strapi build",
-    "strapi": "strapi"
-  },
-  "dependencies": {
-    "@strapi/plugin-i18n": "4.9.1",
-    "@strapi/plugin-users-permissions": "4.9.1",
-    "@strapi/provider-upload-cloudinary": "^4.9.2",
-    "@strapi/strapi": "4.9.1",
-    "cloudinary": "^1.36.1",
-    "pg": "8.8.0"
-  },
-  "author": {
-    "name": "A Strapi developer"
-  },
-  "strapi": {
-    "uuid": "09d0e8a2-0d4e-410a-a3d2-a0cbad319a7a"
-  },
-  "engines": {
-    "node": ">=14.19.1 <=18.x.x",
-    "npm": ">=6.0.0"
-  },
-  "license": "MIT"
-}
