--- strapi/guitarla-strapi/src/api/blog/content-types/blog/schema.json
@@ -1,54 +1,0 @@
-{
-  "kind": "collectionType",
-  "collectionName": "blogs",
-  "info": {
-    "singularName": "blog",
-    "pluralName": "blogs",
-    "displayName": "Blog",
-    "description": ""
-  },
-  "options": {
-    "draftAndPublish": true
-  },
-  "pluginOptions": {},
-  "attributes": {
-    "title": {
-      "type": "string"
-    },
-    "author": {
-      "type": "string"
-    },
-    "publishDate": {
-      "type": "date"
-    },
-    "image": {
-      "type": "media",
-      "multiple": false,
-      "required": true,
-      "allowedTypes": [
-        "images"
-      ]
-    },
-    "url": {
-      "type": "uid",
-      "targetField": "title"
-    },
-    "paragraph": {
-      "type": "text"
-    },
-    "tags": {
-      "type": "string"
-    },
-    "imagen": {
-      "allowedTypes": [
-        "images",
-        "files",
-        "videos",
-        "audios"
-      ],
-      "type": "media",
-      "multiple": false,
-      "required": true
-    }
-  }
-}
