--- strapi/guitarla-strapi/src/api/blog/controllers/blog.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * blog controller
- */
-
-const { createCoreController } = require('@strapi/strapi').factories;
-
-module.exports = createCoreController('api::blog.blog');
