--- strapi/guitarla-strapi/src/api/blog/routes/blog.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * blog router
- */
-
-const { createCoreRouter } = require('@strapi/strapi').factories;
-
-module.exports = createCoreRouter('api::blog.blog');
