--- strapi/guitarla-strapi/src/api/blog/services/blog.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * blog service
- */
-
-const { createCoreService } = require('@strapi/strapi').factories;
-
-module.exports = createCoreService('api::blog.blog');
