--- strapi/guitarla-strapi/src/api/curso/content-types/curso/schema.json
@@ -1,32 +1,0 @@
-{
-  "kind": "singleType",
-  "collectionName": "cursos",
-  "info": {
-    "singularName": "curso",
-    "pluralName": "cursos",
-    "displayName": "Cursos"
-  },
-  "options": {
-    "draftAndPublish": true
-  },
-  "pluginOptions": {},
-  "attributes": {
-    "tituo": {
-      "type": "string",
-      "required": true
-    },
-    "contenido": {
-      "type": "richtext"
-    },
-    "imagen": {
-      "allowedTypes": [
-        "images",
-        "files",
-        "videos",
-        "audios"
-      ],
-      "type": "media",
-      "multiple": false
-    }
-  }
-}
