--- strapi/guitarla-strapi/src/api/curso/controllers/curso.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * curso controller
- */
-
-const { createCoreController } = require('@strapi/strapi').factories;
-
-module.exports = createCoreController('api::curso.curso');
