--- strapi/guitarla-strapi/src/api/curso/routes/curso.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * curso router
- */
-
-const { createCoreRouter } = require('@strapi/strapi').factories;
-
-module.exports = createCoreRouter('api::curso.curso');
