--- strapi/guitarla-strapi/src/api/curso/services/curso.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * curso service
- */
-
-const { createCoreService } = require('@strapi/strapi').factories;
-
-module.exports = createCoreService('api::curso.curso');
