--- strapi/guitarla-strapi/src/api/guitarra/content-types/guitarra/schema.json
@@ -1,38 +1,0 @@
-{
-  "kind": "collectionType",
-  "collectionName": "guitarras",
-  "info": {
-    "singularName": "guitarra",
-    "pluralName": "guitarras",
-    "displayName": "Guitarras",
-    "description": ""
-  },
-  "options": {
-    "draftAndPublish": true
-  },
-  "pluginOptions": {},
-  "attributes": {
-    "nombre": {
-      "type": "string",
-      "required": true
-    },
-    "descripcion": {
-      "type": "richtext"
-    },
-    "precio": {
-      "type": "integer"
-    },
-    "imagen": {
-      "type": "media",
-      "multiple": false,
-      "required": true,
-      "allowedTypes": [
-        "images"
-      ]
-    },
-    "url": {
-      "type": "uid",
-      "targetField": "nombre"
-    }
-  }
-}
