--- strapi/guitarla-strapi/src/api/guitarra/controllers/guitarra.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * guitarra controller
- */
-
-const { createCoreController } = require('@strapi/strapi').factories;
-
-module.exports = createCoreController('api::guitarra.guitarra');
