--- strapi/guitarla-strapi/src/api/guitarra/routes/guitarra.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * guitarra router
- */
-
-const { createCoreRouter } = require('@strapi/strapi').factories;
-
-module.exports = createCoreRouter('api::guitarra.guitarra');
