--- strapi/guitarla-strapi/src/api/guitarra/services/guitarra.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * guitarra service
- */
-
-const { createCoreService } = require('@strapi/strapi').factories;
-
-module.exports = createCoreService('api::guitarra.guitarra');
