--- strapi/guitarla-strapi/src/api/post/content-types/post/schema.json
@@ -1,38 +1,0 @@
-{
-  "kind": "collectionType",
-  "collectionName": "posts",
-  "info": {
-    "singularName": "post",
-    "pluralName": "posts",
-    "displayName": "Post",
-    "description": ""
-  },
-  "options": {
-    "draftAndPublish": true
-  },
-  "pluginOptions": {},
-  "attributes": {
-    "titulo": {
-      "type": "string",
-      "required": true
-    },
-    "contenido": {
-      "type": "richtext"
-    },
-    "imagen": {
-      "type": "media",
-      "multiple": false,
-      "required": false,
-      "allowedTypes": [
-        "images",
-        "files",
-        "videos",
-        "audios"
-      ]
-    },
-    "url": {
-      "type": "uid",
-      "targetField": "titulo"
-    }
-  }
-}
