--- strapi/guitarla-strapi/src/api/post/controllers/post.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * post controller
- */
-
-const { createCoreController } = require('@strapi/strapi').factories;
-
-module.exports = createCoreController('api::post.post');
