--- strapi/guitarla-strapi/src/api/post/routes/post.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * post router
- */
-
-const { createCoreRouter } = require('@strapi/strapi').factories;
-
-module.exports = createCoreRouter('api::post.post');
