--- strapi/guitarla-strapi/src/api/post/services/post.js
@@ -1,9 +1,0 @@
-'use strict';
-
-/**
- * post service
- */
-
-const { createCoreService } = require('@strapi/strapi').factories;
-
-module.exports = createCoreService('api::post.post');
